// src/client/app/shared/models/goodie.ts
export class Goodie {
  winHeadline: string = '';
  winBeschreibung: string = '';
  gType: string = '';
  gTypeText: boolean;
  gTypeCustom: string = '';
  gLegal: string = '';
  gLegalText: boolean;
  gLegalCustom: string = '';
}

// export class GType {
//   gTypeText: boolean;
//   gTypeCustomText: string = '';
// }